Fix broken sign-in
Tonies changed to a KeyCloak-based single-sign-on
The benefit for us is that we now have an actual refresh token and no
longer need to store the user's credentials.

The refresh token is valid for 30 days but that should be enough for
most uses.

This commit also refactors how authentication is handled.